+++ simple_history/models.py
@@ -3,31 +3,23 @@ from __future__ import unicode_literals
 import copy
 import importlib
 import threading
+import uuid
 
+from django.apps import apps
 from django.conf import settings
 from django.contrib import admin
 from django.db import models, router
+from django.db.models import Q
 from django.db.models.fields.proxy import OrderWrt
+from django.urls import reverse
 from django.utils import six
 from django.utils.encoding import python_2_unicode_compatible, smart_text
+from django.utils.text import format_lazy
 from django.utils.timezone import now
-from django.utils.translation import string_concat, ugettext_lazy as _
+from django.utils.translation import ugettext_lazy as _
 
 from . import exceptions
 from .manager import HistoryDescriptor
-
-try:
-    from django.apps import apps
-except ImportError:  # Django < 1.7
-    from django.db.models import get_app
-try:
-    from south.modelsinspector import add_introspection_rules
-except ImportError:  # south not present
-    pass
-else:  # south configuration for CustomForeignKeyField
-    add_introspection_rules(
-        [], ["^simple_history.models.CustomForeignKeyField"])
-
 
 registered_models = {}
 
@@ -37,11 +29,13 @@ class HistoricalRecords(object):
 
     def __init__(self, verbose_name=None, bases=(models.Model,),
                  user_related_name='+', table_name=None, inherit=False,
+                 history_id_field=None,
                  excluded_fields=None):
         self.user_set_verbose_name = verbose_name
         self.user_related_name = user_related_name
         self.table_name = table_name
         self.inherit = inherit
+        self.history_id_field = history_id_field
         if excluded_fields is None:
             excluded_fields = []
         self.excluded_fields = excluded_fields
@@ -72,17 +66,20 @@ class HistoricalRecords(object):
             finally:
                 del self.skip_history_when_saving
             return ret
+
         setattr(cls, 'save_without_historical_record',
                 save_without_historical_record)
 
     def finalize(self, sender, **kwargs):
+        inherited = False
         try:
             hint_class = self.cls
         except AttributeError:  # called via `register`
             pass
         else:
             if hint_class is not sender:  # set in concrete
-                if not (self.inherit and issubclass(sender, hint_class)):
+                inherited = (self.inherit and issubclass(sender, hint_class))
+                if not inherited:
                     return  # set in abstract
         if hasattr(sender._meta, 'simple_history_manager_attribute'):
             raise exceptions.MultipleRegistrationsError(
@@ -91,8 +88,12 @@ class HistoricalRecords(object):
                     sender._meta.object_name,
                 )
             )
-        history_model = self.create_history_model(sender)
-        module = importlib.import_module(self.module)
+        history_model = self.create_history_model(sender, inherited)
+        if inherited:
+            # Make sure history model is in same module as concrete model
+            module = importlib.import_module(history_model.__module__)
+        else:
+            module = importlib.import_module(self.module)
         setattr(module, history_model.__name__, history_model)
 
         # The HistoricalRecords object will be discarded,
@@ -106,7 +107,7 @@ class HistoricalRecords(object):
         setattr(sender, self.manager_name, descriptor)
         sender._meta.simple_history_manager_attribute = self.manager_name
 
-    def create_history_model(self, model):
+    def create_history_model(self, model, inherited):
         """
         Creates a historical model to associate with the model provided.
         """
@@ -116,17 +117,17 @@ class HistoricalRecords(object):
         }
 
         app_module = '%s.models' % model._meta.app_label
-        if model.__module__ != self.module:
+
+        if inherited:
+            # inherited use models module
+            attrs['__module__'] = model.__module__
+        elif model.__module__ != self.module:
             # registered under different app
             attrs['__module__'] = self.module
         elif app_module != self.module:
-            try:
-                # Abuse an internal API because the app registry is loading.
-                app = apps.app_configs[model._meta.app_label]
-            except NameError:  # Django < 1.7
-                models_module = get_app(model._meta.app_label).__name__
-            else:
-                models_module = app.name
+            # Abuse an internal API because the app registry is loading.
+            app = apps.app_configs[model._meta.app_label]
+            models_module = app.name
             attrs['__module__'] = models_module
 
         fields = self.copy_fields(model)
@@ -156,18 +157,15 @@ class HistoricalRecords(object):
         fields = {}
         for field in self.fields_included(model):
             field = copy.copy(field)
-            try:
-                field.remote_field = copy.copy(field.remote_field)
-            except AttributeError:
-                field.rel = copy.copy(field.rel)
+            field.remote_field = copy.copy(field.remote_field)
             if isinstance(field, OrderWrt):
                 # OrderWrt is a proxy field, switch to a plain IntegerField
                 field.__class__ = models.IntegerField
             if isinstance(field, models.ForeignKey):
                 old_field = field
                 field_arguments = {'db_constraint': False}
-                if (getattr(old_field, 'one_to_one', False) or
-                        isinstance(old_field, models.OneToOneField)):
+                if getattr(old_field, 'one_to_one', False) \
+                   or isinstance(old_field, models.OneToOneField):
                     FieldType = models.ForeignKey
                 else:
                     FieldType = type(old_field)
@@ -176,13 +174,15 @@ class HistoricalRecords(object):
                 if getattr(old_field, 'db_column', None):
                     field_arguments['db_column'] = old_field.db_column
 
-                # If old_field.rel.to is 'self' then we have a case where object has a foreign key
-                # to itself. In this case we update need to set the `to` value of the field
-                # to be set to a model. We can use the old_field.model value.
-                if isinstance(old_field.rel.to, str) and old_field.rel.to == 'self':
+                # If old_field.remote_field.model is 'self' then we have a
+                # case where object has a foreign key to itself. In this case
+                # we need to set the `model` value of the field to a model. We
+                # can use the old_field.model value.
+                if isinstance(old_field.remote_field.model, str) and \
+                   old_field.remote_field.model == 'self':
                     object_to = old_field.model
                 else:
-                    object_to = old_field.rel.to
+                    object_to = old_field.remote_field.model
 
                 field = FieldType(
                     object_to,
@@ -207,14 +207,18 @@ class HistoricalRecords(object):
 
         user_model = getattr(settings, 'AUTH_USER_MODEL', 'auth.User')
 
-        @models.permalink
         def revert_url(self):
             """URL for this change in the default admin site."""
             opts = model._meta
             app_label, model_name = opts.app_label, opts.model_name
-            return ('%s:%s_%s_simple_history' %
-                    (admin.site.name, app_label, model_name),
-                    [getattr(self, opts.pk.attname), self.history_id])
+            return reverse(
+                '%s:%s_%s_simple_history' % (
+                    admin.site.name,
+                    app_label,
+                    model_name
+                ),
+                args=[getattr(self, opts.pk.attname), self.history_id]
+            )
 
         def get_instance(self):
             attrs = {
@@ -232,8 +236,35 @@ class HistoricalRecords(object):
                 attrs.update(values)
             return model(**attrs)
 
+        def get_next_record(self):
+            """
+            Get the next history record for the instance. `None` if last.
+            """
+            return self.instance.history.filter(
+                Q(history_date__gt=self.history_date)
+            ).order_by('history_date').first()
+
+        def get_prev_record(self):
+            """
+            Get the previous history record for the instance. `None` if first.
+            """
+            return self.instance.history.filter(
+                Q(history_date__lt=self.history_date)
+            ).order_by('history_date').last()
+
+        if self.history_id_field:
+            history_id_field = self.history_id_field
+            history_id_field.primary_key = True
+            history_id_field.editable = False
+        elif getattr(settings, 'SIMPLE_HISTORY_HISTORY_ID_USE_UUID', False):
+            history_id_field = models.UUIDField(
+                primary_key=True, default=uuid.uuid4, editable=False
+            )
+        else:
+            history_id_field = models.AutoField(primary_key=True)
+
         return {
-            'history_id': models.AutoField(primary_key=True),
+            'history_id': history_id_field,
             'history_date': models.DateTimeField(),
             'history_change_reason': models.CharField(max_length=100,
                                                       null=True),
@@ -245,9 +276,14 @@ class HistoricalRecords(object):
                 ('~', _('Changed')),
                 ('-', _('Deleted')),
             )),
-            'history_object': HistoricalObjectDescriptor(model, self.fields_included(model)),
+            'history_object': HistoricalObjectDescriptor(
+                model,
+                self.fields_included(model)
+            ),
             'instance': property(get_instance),
             'instance_type': model,
+            'next_record': property(get_next_record),
+            'prev_record': property(get_prev_record),
             'revert_url': revert_url,
             '__str__': lambda self: '%s as of %s' % (self.history_object,
                                                      self.history_date)
@@ -265,8 +301,8 @@ class HistoricalRecords(object):
         if self.user_set_verbose_name:
             name = self.user_set_verbose_name
         else:
-            name = string_concat('historical ',
-                                 smart_text(model._meta.verbose_name))
+            name = format_lazy('historical {}',
+                               smart_text(model._meta.verbose_name))
         meta_fields['verbose_name'] = name
         return meta_fields
 
@@ -297,7 +333,7 @@ class HistoricalRecords(object):
             return instance._history_user
         except AttributeError:
             try:
-                if self.thread.request.user.is_authenticated():
+                if self.thread.request.user.is_authenticated:
                     return self.thread.request.user
                 return None
             except AttributeError:
@@ -334,7 +370,8 @@ def convert_auto_field(field):
     must be replaced with an IntegerField.
     """
     connection = router.db_for_write(field.model)
-    if settings.DATABASES[connection].get('ENGINE') in ('django_mongodb_engine',):
+    if settings.DATABASES[connection].get('ENGINE') in \
+       ('django_mongodb_engine',):
         # Check if AutoField is string for django-non-rel support
         return models.TextField
     return models.IntegerField